Move modals one-level up directory
