+++ client/src/components/modals/index.tsx
@@ -1,5 +1,5 @@
-export * from './cancel-order/CancelOrderModal';
-export * from './cancel-profit-order/CancelProfitOrderModal';
-export * from './cancel-all-orders/CancelAllOrdersModal';
-export * from './cancel-all-profit-orders/CancelAllProfitOrdersModal';
-export * from './add-profit-order/AddProfitOrderModal';
+export * from './CancelOrderModal';
+export * from './CancelProfitOrderModal';
+export * from './CancelAllOrdersModal';
+export * from './CancelAllProfitOrdersModal';
+export * from './AddProfitOrderModal';
